Split out individual filters to separate files
With the impending addition of yet another filter, it will get a bit
too confusing to have them all in a single file. This change shouldn't
affect functionality, even if the filters were being accessed
directly.
