+++ filtering/filter/__init__.py
@@ -1,0 +1,2 @@
+from .filter import Filter, FrequencySpaceFilter
+from .spatial_filter import SpatialFilter
